Refactor backend into modular structure
